splitting up cleanup file, cutting down on claudes comments
