+++ apps/examples/src/examples/editor-api/canvas-cleanup/CanvasCleanupExample.tsx
@@ -15,20 +15,14 @@ import './canvas-cleanup.css'
 
 // [1]
 const IDS = {
-	// Group A: text overflow only — spaced apart, but too narrow for their labels
 	userAuth: createShapeId('userAuth'),
 	session: createShapeId('session'),
-	// Group B: overlap only — labels fit, but shapes are piled on top of each other
 	cache: createShapeId('cache'),
 	proxy: createShapeId('proxy'),
-	// Group C: both issues — overlapping AND labels overflow
 	authz: createShapeId('authz'),
 	token: createShapeId('token'),
-	// Group D: no issues — label fits, no overlap (should be untouched)
 	db: createShapeId('db'),
-	// Group E: arrow path only — sits across the arrow, no text or overlap issues
 	middleware: createShapeId('middleware'),
-	// Arrow connecting group A to group C (passes through group E)
 	arrow: createShapeId('arrow'),
 }
 
@@ -91,8 +85,7 @@ function setupCanvas(editor: Editor) {
 				y: 310,
 				props: { w: 130, h: 60, color: 'violet', richText: toRichText('Database') },
 			},
-			// [7] Group E: arrow path only — label and size are fine, no overlap with anything,
-			// but the arrow from group A to group C passes straight through it.
+			// [7] Group E: arrow path only
 			{
 				id: IDS.middleware,
 				type: 'geo',
@@ -133,7 +126,6 @@ function setupCanvas(editor: Editor) {
 	})
 }
 
-// [9]
 function Controls() {
 	const editor = useEditor()
 
@@ -164,7 +156,7 @@ function Controls() {
 export default function CanvasCleanupExample() {
 	return (
 		<div className="tldraw__editor">
-			{/* [10] */}
+			{/* [9] */}
 			<Tldraw
 				onMount={(editor) => {
 					if (editor.getCurrentPageShapeIds().size > 0) return
@@ -178,51 +170,39 @@ export default function CanvasCleanupExample() {
 
 /*
 [1]
-Stable shape IDs let the reset function reliably recreate the same state regardless of
+Stable shape IDs mean the reset function can recreate the same canvas state regardless of
 what the user has done since the last reset.
 
 [2]
-Wrap all shape creation in `editor.run()` so it counts as a single undo step, keeping
-the undo history clean.
+Wrapping all shape creation in `editor.run()` groups it into a single undo step.
 
 [3]
-Group A — text overflow only. The shapes are well-spaced so `resolveShapeOverlaps` won't
-touch them, but their widths are too small for the label words. `resolveTextWordWrap` widens
-each shape and shifts it left so the center stays in place.
+Group A — text overflow only. The shapes are well-spaced but too narrow for their labels.
+`resolveTextWordWrap` widens each shape and shifts it left to keep the center in place.
+`resolveShapeOverlaps` leaves them alone because they don't overlap anything.
 
 [4]
-Group B — overlap only. Labels ("Cache", "Proxy") are short enough to fit within the shape
-width, so `resolveTextWordWrap` leaves them alone. The shapes are deliberately piled on top
-of each other, so `resolveShapeOverlaps` will push them apart.
+Group B — overlap only. The labels fit within the shape width, so `resolveTextWordWrap`
+leaves them alone. The shapes are stacked on top of each other, so `resolveShapeOverlaps`
+pushes them apart.
 
 [5]
-Group C — both issues. The shapes overlap each other AND their labels overflow. Both passes
-will act on this group.
+Group C — both issues. The shapes overlap each other and their labels overflow. Both
+`resolveTextWordWrap` and `resolveShapeOverlaps` act on this group.
 
 [6]
-Group D — no issues. The label fits within the shape width and the shape doesn't overlap
-anything. None of the cleanup passes should move or resize it, demonstrating that clean
-shapes are left untouched.
+Group D — no issues. The label fits and the shape doesn't overlap anything. None of the
+cleanup passes move or resize it.
 
 [7]
 Group E — arrow path only. The label fits and there are no overlaps, but the arrow from
-group A to group C passes straight through this shape. `resolveTextWordWrap` and
-`resolveShapeOverlaps` leave it untouched; only `rerouteArrows` acts on it by adjusting
-the arrow's bend so its path curves around the shape.
+group A to group C passes straight through this shape. Only `rerouteArrows` acts on it,
+adjusting the arrow's bend so its path curves around the shape.
 
 [8]
-The arrow connects a shape from group A to one in group C. Its straight path crosses group E
-(Middleware), which `rerouteArrows` will route around. As the overlap pass moves the endpoint
-shapes, the arrow reroutes automatically — no extra work required for that part.
+The arrow connects a shape in group A to one in group C. Its straight path crosses group E,
+which `rerouteArrows` routes around.
 
 [9]
-The Controls component is rendered via the `TopPanel` slot so it sits inside the canvas.
-Each button calls one of the four cleanup utilities:
-- `resolveTextWordWrap`: widens shapes whose labels would break a word across lines.
-- `resolveShapeOverlaps`: pushes overlapping shapes apart until each pair has ≥ 20px of space.
-- `rerouteArrows`: adjusts arrow bend values so paths avoid non-endpoint bystander shapes.
-- `cleanupCanvas`: runs all three in order as a single undoable step.
-
-[10]
 Guard against re-running setup when the canvas already has shapes (e.g. after a hot-reload).
 */
